Restructure classes a bit
