+++ classes/suggested-tasks/local-tasks/providers/content/class-content-abstract.php
@@ -5,8 +5,9 @@
  * @package Progress_Planner
  */
 
-namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Content;
 
+use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Local_Tasks_Abstract;
 /**
  * Add tasks for content updates.
  */
