+++ classes/suggested-tasks/local-tasks/providers/one-time/class-blog-description.php
@@ -1,23 +1,23 @@
 <?php
 /**
- * Add tasks for Core siteicon.
+ * Add tasks for Core blogdescription.
  *
  * @package Progress_Planner
  */
 
-namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\One_Time;
 
 /**
- * Add tasks for Core siteicon.
+ * Add tasks for Core blogdescription.
  */
-class Core_Siteicon extends Local_OneTime_Tasks_Abstract {
+class Blog_Description extends One_Time {
 
 	/**
 	 * The provider ID.
 	 *
 	 * @var string
 	 */
-	const ID = 'core-siteicon';
+	const ID = 'core-blogdescription';
 
 	/**
 	 * The provider type.
@@ -32,8 +32,7 @@ class Core_Siteicon extends Local_OneTime_Tasks_Abstract {
 	 * @return bool
 	 */
 	public function should_add_task() {
-		$site_icon = \get_option( 'site_icon' );
-		return '' === $site_icon || '0' === $site_icon;
+		return '' === \get_bloginfo( 'description' );
 	}
 
 	/**
@@ -51,16 +50,16 @@ class Core_Siteicon extends Local_OneTime_Tasks_Abstract {
 
 		return [
 			'task_id'     => $task_id,
-			'title'       => \esc_html__( 'Set site icon', 'progress-planner' ),
+			'title'       => \esc_html__( 'Set tagline', 'progress-planner' ),
 			'parent'      => 0,
 			'priority'    => 'high',
 			'type'        => $this->get_provider_type(),
 			'points'      => 1,
 			'url'         => $this->capability_required() ? \esc_url( \admin_url( 'options-general.php' ) ) : '',
 			'description' => '<p>' . sprintf(
-				/* translators: %s:<a href="https://prpl.fyi/set-site-icon" target="_blank">site icon</a> link */
+				/* translators: %s:<a href="https://prpl.fyi/set-tagline" target="_blank">tagline</a> link */
 				\esc_html__( 'Set the %s to make your website look more professional.', 'progress-planner' ),
-				'<a href="https://prpl.fyi/set-site-icon" target="_blank">' . \esc_html__( 'site icon', 'progress-planner' ) . '</a>'
+				'<a href="https://prpl.fyi/set-tagline" target="_blank">' . \esc_html__( 'tagline', 'progress-planner' ) . '</a>'
 			) . '</p>',
 		];
 	}
