+++ classes/suggested-tasks/local-tasks/providers/one-time/class-one-time.php
@@ -5,14 +5,14 @@
  * @package Progress_Planner
  */
 
-namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\One_Time;
 
 use Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\Local_Tasks_Abstract;
-use Progress_Planner\Suggested_Tasks\Local_Tasks\Local_Task_Factory;
+
 /**
  * Add tasks for content updates.
  */
-abstract class Local_Repetitive_Tasks_Abstract extends Local_Tasks_Abstract {
+abstract class One_Time extends Local_Tasks_Abstract {
 
 	/**
 	 * Evaluate a task.
@@ -28,14 +28,7 @@ abstract class Local_Repetitive_Tasks_Abstract extends Local_Tasks_Abstract {
 			return false;
 		}
 
-		$task_object = ( new Local_Task_Factory( $task_id ) )->get_task();
-		$task_data   = $task_object->get_data();
-
-		if ( $task_data['type'] === $this->get_provider_id() && \gmdate( 'YW' ) === $task_data['year_week'] && $this->is_task_completed() ) {
-			return $task_id;
-		}
-
-		return false;
+		return $this->is_task_completed() ? $task_id : false;
 	}
 
 	/**
@@ -70,19 +63,16 @@ abstract class Local_Repetitive_Tasks_Abstract extends Local_Tasks_Abstract {
 	 * @return array
 	 */
 	public function get_tasks_to_inject() {
-
-		$task_id = $this->get_provider_id() . '-' . \gmdate( 'YW' );
-
 		if (
 			true === $this->is_task_type_snoozed() ||
 			! $this->should_add_task() || // No need to add the task.
-			true === \progress_planner()->get_suggested_tasks()->was_task_completed( $task_id )
+			true === \progress_planner()->get_suggested_tasks()->was_task_completed( $this->get_provider_id() )
 		) {
 			return [];
 		}
 
 		return [
-			$this->get_task_details( $task_id ),
+			$this->get_task_details( $this->get_provider_id() ),
 		];
 	}
 }
