+++ classes/suggested-tasks/local-tasks/providers/one-time/class-remove-inactive-plugins.php
@@ -5,12 +5,12 @@
  * @package Progress_Planner
  */
 
-namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\One_Time;
 
 /**
  * Add tasks to check if WP debug is enabled.
  */
-class Remove_Inactive_Plugins extends Local_OneTime_Tasks_Abstract {
+class Remove_Inactive_Plugins extends One_Time {
 
 	/**
 	 * The provider type.
