+++ classes/suggested-tasks/local-tasks/providers/one-time/class-site-icon.php
@@ -1,23 +1,23 @@
 <?php
 /**
- * Add tasks for Core blogdescription.
+ * Add tasks for Core siteicon.
  *
  * @package Progress_Planner
  */
 
-namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers\One_Time;
 
 /**
- * Add tasks for Core blogdescription.
+ * Add tasks for Core siteicon.
  */
-class Core_Blogdescription extends Local_OneTime_Tasks_Abstract {
+class Site_Icon extends One_Time {
 
 	/**
 	 * The provider ID.
 	 *
 	 * @var string
 	 */
-	const ID = 'core-blogdescription';
+	const ID = 'core-siteicon';
 
 	/**
 	 * The provider type.
@@ -32,7 +32,8 @@ class Core_Blogdescription extends Local_OneTime_Tasks_Abstract {
 	 * @return bool
 	 */
 	public function should_add_task() {
-		return '' === \get_bloginfo( 'description' );
+		$site_icon = \get_option( 'site_icon' );
+		return '' === $site_icon || '0' === $site_icon;
 	}
 
 	/**
@@ -50,16 +51,16 @@ class Core_Blogdescription extends Local_OneTime_Tasks_Abstract {
 
 		return [
 			'task_id'     => $task_id,
-			'title'       => \esc_html__( 'Set tagline', 'progress-planner' ),
+			'title'       => \esc_html__( 'Set site icon', 'progress-planner' ),
 			'parent'      => 0,
 			'priority'    => 'high',
 			'type'        => $this->get_provider_type(),
 			'points'      => 1,
 			'url'         => $this->capability_required() ? \esc_url( \admin_url( 'options-general.php' ) ) : '',
 			'description' => '<p>' . sprintf(
-				/* translators: %s:<a href="https://prpl.fyi/set-tagline" target="_blank">tagline</a> link */
+				/* translators: %s:<a href="https://prpl.fyi/set-site-icon" target="_blank">site icon</a> link */
 				\esc_html__( 'Set the %s to make your website look more professional.', 'progress-planner' ),
-				'<a href="https://prpl.fyi/set-tagline" target="_blank">' . \esc_html__( 'tagline', 'progress-planner' ) . '</a>'
+				'<a href="https://prpl.fyi/set-site-icon" target="_blank">' . \esc_html__( 'site icon', 'progress-planner' ) . '</a>'
 			) . '</p>',
 		];
 	}
